+++ src/__debug__.ts
@@ -3,12 +3,12 @@ import {SExp} from "./SExp";
 export function prettyPrint(enable: boolean){
   if(enable){
     const toString = SExp.prototype.toString;
-    SExp.prototype.toString = SExp.prototype.__repl__;
-    SExp.prototype.__repl__ = toString;
+    SExp.prototype.toString = SExp.prototype.__repr__;
+    SExp.prototype.__repr__ = toString;
   }
   else{
-    const __repl__ = SExp.prototype.toString;
-    SExp.prototype.toString = SExp.prototype.__repl__;
-    SExp.prototype.__repl__ = __repl__;
+    const __repr__ = SExp.prototype.toString;
+    SExp.prototype.toString = SExp.prototype.__repr__;
+    SExp.prototype.__repr__ = __repr__;
   }
 }